// src/entities/countryInput/countryInput.types.ts
export interface ICountryInputData {
  updated: string
  country: string
  countryInfo: {
    _id: number
    iso2: string
    iso3: string
    lat: number
    long: number
    flag: string
  }
  cases: number
  todayCases: number
  deaths: number
  todayDeaths: number
  recovered: number
  todayRecovered: number
  active: number
  critical: number
  casesPerOneMillion: number
  deathsPerOneMillion: number
  tests: number
  testsPerOneMillion: number
  population: number
  continent: string
  oneCasePerPeople: number
  oneDeathPerPeople: number
  oneTestPerPeople: number
  activePerOneMillion: number
  recoveredPerOneMillion: number
  criticalPerOneMillion: number
}

export interface ICountryInput extends ICountryInputData{

}

export interface ICountryOptionData {
  countryCode: string
  country: string
}

export interface ICountryOption extends ICountryOptionData{

}